requirements.txt
tensorflow
opencv-python
Keras
pillow
matplotlib
scikit-image
flask
waitress